Split control_loop.py into runner/reduce/streams package
Pure mechanical split, no logic changes. control_loop becomes a package
re-exporting the prior public + test-facing surface so all import sites
keep working. The one streams->reduce edge (_add_or_enqueue_event in the
release-firing path) is an inline import at that chokepoint.
